Move TestNamedayCalendarBuilder to the resource package
